refactor(ai): extract ai_completion_client and streamline test runner
